requirements/dev.txt
black==24.1.1
flake8==7.0.0
isort==5.13.2
pytest==8.0.0
pytest-django==4.8.0
django-debug-toolbar==4.3.0
